Addressing CR comments and some refactoring
- Use the same helper function when loading keys from keystore files when
signing targets as when creating a new repository
- Enabled users to keep entering key pem until valid
- Moved functions used to load keys from keystore files and from the cmd into a separate module as a first step towars refactoring developer_tool
- Moved one function to yubikey module
